+++ jest.config.js
@@ -22,6 +22,7 @@ module.exports = {
     "src/types/index.ts",
     "src/utils/ops-stack-api-utils/get-client.ts",
     "src/utils/ops-stack-api-utils/index.ts",
-    "src/utils/ops-core/index.ts"
+    "src/utils/ops-core/index.ts",
+    "src/utils/spinner/index.ts"
   ]
 };