+++ test/commands/deploy.test.ts
@@ -9,6 +9,10 @@ const mockCreateOpsStack = jest.fn();
 const mockUpdateOpsStack = jest.fn();
 const mockGetCredentials = jest.fn();
 const mockGetClient = jest.fn();
+const mockGetSpinner = jest.fn();
+const mockStart = jest.fn();
+const mockStop = jest.fn();
+const mockSleep = jest.fn();
 
 jest.mock('fs', () => ({
   readFileSync: mockReadFileSync
@@ -37,13 +41,28 @@ jest.mock('../../src/utils/ops-stack-api-utils', () => ({
   getClient: mockGetClient
 }));
 
-import { deploy } from '../../src/commands/deploy';
+jest.mock('../../src/utils/spinner', () => ({
+  getSpinner: mockGetSpinner
+}));
 
+jest.mock('../../src/utils/os', () => ({
+  sleep: mockSleep
+}));
+
+import HttpError from 'http-errors';
+import { deploy } from '../../src/commands/deploy';
 
 const mockConfigYaml = 'Console: name: console';
 const mockConfigJson = { name: 'console' };
 const mockOpsStack = { name: 'console', status: 'Syncing' };
 const mockCredentials = { apiKey: 'mock-api-key' };
+
+const mockSpinner = {
+  text: '',
+  color: '',
+  start: mockStart,
+  stop: mockStop
+}
 
 const mockOpsStackApiClient = {
   allocate: {
@@ -55,8 +74,10 @@ const mockOpsStackApiClient = {
 
 describe('deploy', () => {
   beforeEach(() => {
+    mockSleep.mockResolvedValue(undefined);
     mockGetClient.mockReturnValue(mockOpsStackApiClient);
     mockGetCredentials.mockReturnValue(mockCredentials);
+    mockGetSpinner.mockResolvedValue({ ...mockSpinner });
   });
   afterEach(() => {
     // for mocks
@@ -69,6 +90,7 @@ describe('deploy', () => {
     mockReadFileSync.mockReturnValueOnce(mockConfigYaml);
     mockParseConfig.mockResolvedValueOnce(mockConfigJson);
     mockGetOpsStack.mockRejectedValueOnce({ status: 404 });
+    mockGetOpsStack.mockResolvedValueOnce([{ ...mockOpsStack, status: 'In Sync' }]);
     mockCreateOpsStack.mockResolvedValueOnce(mockOpsStack);
 
     await deploy({ configFile: 'index.yml' });
@@ -84,11 +106,23 @@ describe('deploy', () => {
 
     expect(mockGetClient).toBeCalled();
 
+    expect(mockGetOpsStack).toBeCalled();
+    expect(mockGetOpsStack).toBeCalledTimes(2);
+
     expect(mockCreateOpsStack).toBeCalled();
     expect(mockCreateOpsStack).toBeCalledWith(mockConfigYaml);
 
+    expect(mockGetSpinner).toBeCalled();
+    expect(mockGetSpinner).toBeCalledWith({
+      text: 'Initializing...',
+      color: 'blue'
+    });
+
+    expect(mockStart).toBeCalled();
+    expect(mockStop).toBeCalled();
+
     expect(mockLoggerSuccess).toBeCalled();
-    expect(mockLoggerSuccess).toBeCalledWith('Successful started ops console deployment!');
+    expect(mockLoggerSuccess).toBeCalledWith('Successfully deployed ops console!');
     
     expect(mockLoggerStdout).toBeCalled();
     expect(mockLoggerStdout).toBeCalledWith(JSON.stringify(mockOpsStack, null, 2));
@@ -99,6 +133,7 @@ describe('deploy', () => {
     mockReadFileSync.mockReturnValueOnce(mockConfigYaml);
     mockParseConfig.mockResolvedValueOnce(mockConfigJson);
     mockGetOpsStack.mockResolvedValueOnce([mockOpsStack]);
+    mockGetOpsStack.mockResolvedValueOnce([{ ...mockOpsStack, status: 'In Sync' }]);
     mockUpdateOpsStack.mockResolvedValueOnce(mockOpsStack);
 
     await deploy({});
@@ -119,8 +154,17 @@ describe('deploy', () => {
     expect(mockUpdateOpsStack).toBeCalled();
     expect(mockUpdateOpsStack).toBeCalledWith('console', mockConfigYaml);
 
+    expect(mockGetSpinner).toBeCalled();
+    expect(mockGetSpinner).toBeCalledWith({
+      text: 'Initializing...',
+      color: 'blue'
+    });
+
+    expect(mockStart).toBeCalled();
+    expect(mockStop).toBeCalled();
+
     expect(mockLoggerSuccess).toBeCalled();
-    expect(mockLoggerSuccess).toBeCalledWith('Successful started ops console deployment!');
+    expect(mockLoggerSuccess).toBeCalledWith('Successfully deployed ops console!');
     
     expect(mockLoggerStdout).toBeCalled();
     expect(mockLoggerStdout).toBeCalledWith(JSON.stringify(mockOpsStack, null, 2));
@@ -141,7 +185,127 @@ describe('deploy', () => {
     expect(mockLoggerSuccess).not.toBeCalled();    
     expect(mockLoggerStdout).not.toBeCalled();
 
+    expect(mockStop).toBeCalled();
     expect(mockLoggerError).toBeCalled();
     expect(mockLoggerError).toBeCalledWith(`Error deploying ops console: ${mockError}`);
+  });
+  it('throws if the ops stack create/update call fails', async () => {
+    const mockError = HttpError.InternalServerError('Error!');
+    mockResolve.mockImplementationOnce((filePath: string) => filePath);
+    mockReadFileSync.mockReturnValueOnce(mockConfigYaml);
+    mockParseConfig.mockResolvedValueOnce(mockConfigJson);
+    mockGetOpsStack.mockRejectedValueOnce({ status: 404 });
+    mockCreateOpsStack.mockResolvedValueOnce(mockError);
+
+    await deploy({ configFile: 'index.yml' });
+
+    expect(mockGetSpinner).toBeCalled();
+    expect(mockGetSpinner).toBeCalledWith({
+      text: 'Initializing...',
+      color: 'blue'
+    });
+
+    expect(mockResolve).toBeCalled()
+    expect(mockResolve).toBeCalledWith('index.yml');
+
+    expect(mockReadFileSync).toBeCalled();
+    expect(mockReadFileSync).toBeCalledWith('index.yml');
+    
+    expect(mockParseConfig).toBeCalled();
+    expect(mockParseConfig).toBeCalledWith('index.yml');
+
+    expect(mockGetClient).toBeCalled();
+
+    expect(mockGetOpsStack).toBeCalled();
+    expect(mockGetOpsStack).toBeCalledTimes(1);
+
+    expect(mockStart).toBeCalled();
+
+    expect(mockCreateOpsStack).toBeCalled();
+    expect(mockCreateOpsStack).toBeCalledWith(mockConfigYaml);
+
+    expect(mockStop).toBeCalled();
+
+    expect(mockLoggerError).toBeCalled();
+    expect(mockLoggerError).toBeCalledWith(`Error deploying ops console: ${mockError}`);
+  });
+  it('throws if the ops stack has a status of Sync Failed', async () => {
+    mockResolve.mockImplementationOnce((filePath: string) => filePath);
+    mockReadFileSync.mockReturnValueOnce(mockConfigYaml);
+    mockParseConfig.mockResolvedValueOnce(mockConfigJson);
+    mockGetOpsStack.mockRejectedValueOnce({ status: 404 });
+    mockGetOpsStack.mockResolvedValueOnce([{ ...mockOpsStack, status: 'Sync Failed' }]);
+    mockCreateOpsStack.mockResolvedValueOnce(mockOpsStack);
+
+    await deploy({ configFile: 'index.yml' });
+
+    expect(mockResolve).toBeCalled()
+    expect(mockResolve).toBeCalledWith('index.yml');
+
+    expect(mockReadFileSync).toBeCalled();
+    expect(mockReadFileSync).toBeCalledWith('index.yml');
+    
+    expect(mockParseConfig).toBeCalled();
+    expect(mockParseConfig).toBeCalledWith('index.yml');
+
+    expect(mockGetClient).toBeCalled();
+
+    expect(mockGetOpsStack).toBeCalled();
+    expect(mockGetOpsStack).toBeCalledTimes(2);
+
+    expect(mockCreateOpsStack).toBeCalled();
+    expect(mockCreateOpsStack).toBeCalledWith(mockConfigYaml);
+
+    expect(mockGetSpinner).toBeCalled();
+    expect(mockGetSpinner).toBeCalledWith({
+      text: 'Initializing...',
+      color: 'blue'
+    });
+
+    expect(mockStart).toBeCalled();
+    expect(mockStop).toBeCalled();
+
+    expect(mockLoggerError).toBeCalled();
+    expect(mockLoggerError).toBeCalledWith(`Error deploying ops console: ${new Error('Failed to deploy ops console!')}`);
+  });
+  it('throws if the get ops stack call fails', async () => {
+    const mockError = new Error('Error!');
+    mockResolve.mockImplementationOnce((filePath: string) => filePath);
+    mockReadFileSync.mockReturnValueOnce(mockConfigYaml);
+    mockParseConfig.mockResolvedValueOnce(mockConfigJson);
+    mockGetOpsStack.mockRejectedValueOnce({ status: 404 });
+    mockGetOpsStack.mockRejectedValueOnce(mockError);
+    mockCreateOpsStack.mockResolvedValueOnce(mockOpsStack);
+
+    await deploy({ configFile: 'index.yml' });
+
+    expect(mockResolve).toBeCalled()
+    expect(mockResolve).toBeCalledWith('index.yml');
+
+    expect(mockReadFileSync).toBeCalled();
+    expect(mockReadFileSync).toBeCalledWith('index.yml');
+    
+    expect(mockParseConfig).toBeCalled();
+    expect(mockParseConfig).toBeCalledWith('index.yml');
+
+    expect(mockGetClient).toBeCalled();
+
+    expect(mockGetOpsStack).toBeCalled();
+    expect(mockGetOpsStack).toBeCalledTimes(2);
+
+    expect(mockCreateOpsStack).toBeCalled();
+    expect(mockCreateOpsStack).toBeCalledWith(mockConfigYaml);
+
+    expect(mockGetSpinner).toBeCalled();
+    expect(mockGetSpinner).toBeCalledWith({
+      text: 'Initializing...',
+      color: 'blue'
+    });
+
+    expect(mockStart).toBeCalled();
+    expect(mockStop).toBeCalled();
+
+    expect(mockLoggerError).toBeCalled();
+    expect(mockLoggerError).toBeCalledWith(`Error deploying ops console: ${new Error('Error!')}`);
   });
 });